RustC: Move to own translation unit
